tests/test_annotate.py
"""Tests for MapAnnotator class."""
import pytest
import json
import yaml
import shutil
from pathlib import Path
from PIL import Image
from unittest.mock import Mock, patch, MagicMock


class TestMapAnnotatorInit:
    """Tests for MapAnnotator initialization."""

    def test_init_requires_config_file(self):
        """Test that MapAnnotator requires config.yaml file."""
        # TODO: Current implementation doesn't handle missing config file
        # Should raise helpful error message
        with pytest.raises(FileNotFoundError):
            from annotate import MapAnnotator
            # This will fail if data/config.yaml doesn't exist
            annotator = MapAnnotator()

    @patch('annotate.inspect')
    def test_init_with_config(self, mock_inspect, temp_dir, sample_config, sample_marks_data, sample_zone_info):
        """Test MapAnnotator initialization with valid config."""
        # TODO: This test is complex due to file path dependencies
        # Need to mock file system or create temporary structure
        pytest.skip("Requires complex setup with data directory structure")

    def test_init_loads_config_with_utf8(self):
        """Test that config is loaded with UTF-8 encoding."""
        # TODO: Test that config file is opened with encoding="utf-8"
        # This verifies the fix for encoding issues
        pytest.skip("Requires mocking file open to verify encoding parameter")


class TestMapAnnotatorPaths:
    """Tests for path generation methods."""

    @pytest.fixture
    def mock_annotator(self, temp_dir):
        """Create a mock MapAnnotator with minimal setup."""
        # TODO: Need to create proper mock or fixture
        # This is complex due to initialization dependencies
        pytest.skip("Requires full MapAnnotator setup")

    def test_get_path_basic(self, mock_annotator):
        """Test basic path generation."""
        pytest.skip("Requires mock_annotator fixture")

    def test_get_path_with_backup(self, mock_annotator):
        """Test path generation for backup files."""
        pytest.skip("Requires mock_annotator fixture")

    def test_get_path_with_project(self, mock_annotator):
        """Test path generation for project files."""
        pytest.skip("Requires mock_annotator fixture")


class TestMapAnnotatorFileOps:
    """Tests for file operation methods."""

    def test_check_files_all_exist(self):
        """Test check_files when all files exist."""
        # TODO: Need to setup proper directory structure
        pytest.skip("Requires full file structure setup")

    def test_check_files_missing_files(self):
        """Test check_files when files are missing."""
        # TODO: Should print missing files, not raise error
        pytest.skip("Requires full file structure setup")

    def test_backup_files_with_warning(self):
        """Test backup_files with warning enabled."""
        # TODO: Should print warning and return without action
        pytest.skip("Requires full file structure setup")

    def test_backup_files_without_warning(self):
        """Test backup_files actually copies files."""
        # TODO: Should copy files to backup location
        pytest.skip("Requires full file structure setup")


class TestMapAnnotatorZoneMarks:
    """Tests for zone and mark handling."""

    def test_get_zone_marks_basic(self):
        """Test getting marks for a zone."""
        pytest.skip("Requires MapAnnotator setup")

    def test_get_zone_marks_with_rank_remap(self):
        """Test rank remapping for multiple marks of same rank."""
        pytest.skip("Requires MapAnnotator setup")

    def test_check_spawn_points_no_overlap(self):
        """Test checking spawn points with no overlaps."""
        pytest.skip("Requires MapAnnotator setup")

    def test_check_spawn_points_with_overlap(self):
        """Test checking spawn points with overlapping positions."""
        pytest.skip("Requires MapAnnotator setup")

    def test_check_spawn_points_custom_threshold(self):
        """Test checking spawn points with custom threshold."""
        pytest.skip("Requires MapAnnotator setup")


class TestMapAnnotatorAnnotation:
    """Tests for map annotation methods."""

    def test_annotate_map_basic(self):
        """Test basic map annotation."""
        # TODO: Requires:
        # - Mock image loading
        # - Mock file system
        # - Mock ImageMagick
        pytest.skip("Complex test requiring extensive mocking")

    def test_annotate_map_with_save(self):
        """Test annotating and saving map."""
        pytest.skip("Requires full setup with ImageMagick")

    def test_annotate_map_with_unicode_zone(self):
        """Test annotating map with Unicode zone name."""
        pytest.skip("Requires full setup")

    def test_annotate_all_processes_all_zones(self):
        """Test that annotate_all processes all zones."""
        pytest.skip("Requires full setup")

    def test_draw_marker_rank_s(self):
        """Test drawing S rank marker."""
        pytest.skip("Requires MapAnnotator instance")

    def test_draw_marker_rank_a(self):
        """Test drawing A rank markers."""
        pytest.skip("Requires MapAnnotator instance")

    def test_draw_marker_rank_b(self):
        """Test drawing B rank markers."""
        pytest.skip("Requires MapAnnotator instance")

    def test_draw_marker_rank_ss(self):
        """Test drawing SS rank marker."""
        pytest.skip("Requires MapAnnotator instance")


class TestMapAnnotatorSave:
    """Tests for map saving functionality."""

    def test_save_map_requires_imagemagick(self):
        """Test that saving map requires ImageMagick."""
        # TODO: Should provide helpful error if ImageMagick not found
        pytest.skip("Requires ImageMagick setup")

    def test_save_map_handles_subprocess_error(self):
        """Test handling of ImageMagick subprocess errors."""
        # TODO: Current implementation doesn't check subprocess return code
        # subprocess.run should check for errors
        pytest.skip("Requires mocking subprocess")

    def test_save_map_creates_directories(self):
        """Test that save_map creates necessary directories."""
        pytest.skip("Requires full setup")

    def test_save_map_creates_preview(self):
        """Test that save_map creates PNG preview."""
        pytest.skip("Requires full setup")


class TestMapAnnotatorBlend:
    """Tests for map blending functionality."""

    def test_blend_map_basic(self):
        """Test basic map blending."""
        pytest.skip("Requires full setup with mask files")

    def test_blend_map_size_mismatch_raises_error(self):
        """Test that size mismatch raises ValueError."""
        # TODO: Current implementation raises ValueError without message
        # Should provide helpful error message
        pytest.skip("Requires full setup")

    def test_blend_map_missing_mask_file(self):
        """Test handling of missing mask file."""
        # TODO: Should provide helpful error message
        pytest.skip("Requires full setup")

    def test_blend_all_processes_all_zones(self):
        """Test that blend_all processes all zones."""
        pytest.skip("Requires full setup")


class TestMapAnnotatorThumbnail:
    """Tests for thumbnail generation."""

    def test_generate_thumbnail_table_basic(self):
        """Test generating thumbnail table HTML."""
        pytest.skip("Requires full setup")

    def test_generate_thumbnail_table_copies_to_clipboard(self):
        """Test that generated HTML is copied to clipboard."""
        pytest.skip("Requires pyperclip mocking")

    def test_generate_thumbnail_table_handles_norvrandt(self):
        """Test special handling of Norvrandt region."""
        pytest.skip("Requires full setup")


class TestIntegrationAnnotate:
    """Integration tests for annotation workflow."""

    def test_full_annotation_workflow(self):
        """Test complete workflow: load config, annotate, save."""
        # TODO: This would be a full end-to-end test
        # Requires:
        # - Sample config, marks, zones
        # - Sample images
        # - ImageMagick
        pytest.skip("Complex integration test")

    def test_annotation_with_unicode_data(self):
        """Test annotation workflow with Unicode zone/mark names."""
        pytest.skip("Complex integration test")

    def test_annotation_error_recovery(self):
        """Test that errors in one zone don't break others."""
        # TODO: Current implementation may fail on first error
        pytest.skip("Requires error handling improvements")


class TestYAMLSafeLoading:
    """Tests for YAML loading security."""

    def test_config_uses_safe_loader(self):
        """Test that config loading uses safe YAML loader."""
        # TODO: Current implementation uses yaml.Loader instead of SafeLoader
        # This is a security concern
        pytest.skip("Need to verify YAML loader type in actual code")

    def test_zone_info_uses_safe_loader(self):
        """Test that zone info loading uses safe YAML loader."""
        # TODO: Current implementation uses yaml.Loader
        pytest.skip("Need to verify YAML loader type in actual code")
