pyproject.toml
[project]
name = "ffxiv-huntmaps-maker"
version = "0.1.0"
description = "Library + CLI to annotate FFXIV in-game map assets with Elite Marks spawn positions"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Arkhelyi" }
]
keywords = ["ffxiv", "maps", "elite marks"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "fire",
    "numpy",
    "pillow",
    "pyperclip",
    "pyyaml",
    "requests",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "responses>=0.23.0",
]

[project.scripts]
ffxiv-huntmaps = "annotate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
